+++ checkpoint/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
- - Add `metadata()` method for loading with SafeTensors checkpoint format.
+- Add `metadata()` method for loading with SafeTensors checkpoint format.
+- Add options to control number of replicas and min bytes in replica-parallel.
 
 ## [0.11.20] - 2025-07-28